combine into one class
